Parallel SDK testing (-stabilize parallel tests)
