+++ pkg/networkservice/common/onidle/server_test.go
@@ -40,8 +40,7 @@ const (
 )
 
 func TestIdleNotifier_Disable(t *testing.T) {
-	t.Parallel()
-	t.Cleanup(func() { goleak.VerifyNone(t, goleak.IgnoreCurrent()) })
+	t.Cleanup(func() { goleak.VerifyNone(t) })
 
 	ctx, cancel := context.WithTimeout(context.Background(), time.Second)
 	defer cancel()
@@ -58,8 +57,7 @@ func TestIdleNotifier_Disable(t *testing.T) {
 }
 
 func TestIdleNotifier_NoRequests(t *testing.T) {
-	t.Parallel()
-	t.Cleanup(func() { goleak.VerifyNone(t, goleak.IgnoreCurrent()) })
+	t.Cleanup(func() { goleak.VerifyNone(t) })
 
 	ctx, cancel := context.WithTimeout(context.Background(), time.Second)
 	defer cancel()
@@ -80,8 +78,7 @@ func TestIdleNotifier_NoRequests(t *testing.T) {
 }
 
 func TestIdleNotifier_Refresh(t *testing.T) {
-	t.Parallel()
-	t.Cleanup(func() { goleak.VerifyNone(t, goleak.IgnoreCurrent()) })
+	t.Cleanup(func() { goleak.VerifyNone(t) })
 
 	ctx, cancel := context.WithTimeout(context.Background(), time.Second)
 	defer cancel()
@@ -109,8 +106,7 @@ func TestIdleNotifier_Refresh(t *testing.T) {
 }
 
 func TestIdleNotifier_HoldingActiveRequest(t *testing.T) {
-	t.Parallel()
-	t.Cleanup(func() { goleak.VerifyNone(t, goleak.IgnoreCurrent()) })
+	t.Cleanup(func() { goleak.VerifyNone(t) })
 
 	ctx, cancel := context.WithTimeout(context.Background(), time.Second)
 	defer cancel()
@@ -148,8 +144,7 @@ func TestIdleNotifier_HoldingActiveRequest(t *testing.T) {
 }
 
 func TestIdleNotifier_FailedRequest(t *testing.T) {
-	t.Parallel()
-	t.Cleanup(func() { goleak.VerifyNone(t, goleak.IgnoreCurrent()) })
+	t.Cleanup(func() { goleak.VerifyNone(t) })
 
 	ctx, cancel := context.WithTimeout(context.Background(), time.Second)
 	defer cancel()
@@ -178,8 +173,7 @@ func TestIdleNotifier_FailedRequest(t *testing.T) {
 }
 
 func TestIdleNotifier_ContextCancel(t *testing.T) {
-	t.Parallel()
-	t.Cleanup(func() { goleak.VerifyNone(t, goleak.IgnoreCurrent()) })
+	t.Cleanup(func() { goleak.VerifyNone(t) })
 
 	ctx, cancel := context.WithTimeout(context.Background(), time.Second)
 	defer cancel()
@@ -201,8 +195,7 @@ func TestIdleNotifier_ContextCancel(t *testing.T) {
 }
 
 func TestIdleNotifier_RequestAfterExpire(t *testing.T) {
-	t.Parallel()
-	t.Cleanup(func() { goleak.VerifyNone(t, goleak.IgnoreCurrent()) })
+	t.Cleanup(func() { goleak.VerifyNone(t) })
 
 	ctx, cancel := context.WithTimeout(context.Background(), time.Second)
 	defer cancel()
