+++ pkg/registry/common/connect/ns_server_test.go
@@ -244,8 +244,7 @@ func TestConnectNSServer_AllDead_WatchingFind(t *testing.T) {
 	}
 }
 
-func Test_NSConenctChain_Find(t *testing.T) {
-	t.Parallel()
+func Test_NSConnectChain_Find(t *testing.T) {
 	for depth := 2; depth < 11; depth++ {
 		for killIndex := 1; killIndex < depth; killIndex++ {
 			var ctx, cancel = context.WithTimeout(context.Background(), time.Second)
