+++ pkg/tools/clockmock/mock_test.go
@@ -34,123 +34,8 @@ const (
 	testTick = testWait / 100
 )
 
-func TestMock_SetSpeed(t *testing.T) {
-	t.Parallel()
-	samples := []struct {
-		name                    string
-		firstSpeed, secondSpeed float64
-	}{
-		{
-			name:        "From 0",
-			firstSpeed:  0,
-			secondSpeed: 1,
-		},
-		{
-			name:        "To 0",
-			firstSpeed:  1,
-			secondSpeed: 0,
-		},
-		{
-			name:        "Same",
-			firstSpeed:  1,
-			secondSpeed: 1,
-		},
-		{
-			name:        "Increasing to",
-			firstSpeed:  0.1,
-			secondSpeed: 1,
-		},
-		{
-			name:        "Increasing from",
-			firstSpeed:  1,
-			secondSpeed: 10,
-		},
-		{
-			name:        "Decreasing to",
-			firstSpeed:  10,
-			secondSpeed: 1,
-		},
-		{
-			name:        "Decreasing from",
-			firstSpeed:  1,
-			secondSpeed: 0.1,
-		},
-	}
-
-	speeds := []struct {
-		name       string
-		multiplier float64
-	}{
-		{
-			name:       "Slow",
-			multiplier: 0.001,
-		},
-		{
-			name:       "Real",
-			multiplier: 1,
-		},
-		{
-			name:       "Fast",
-			multiplier: 1000,
-		},
-	}
-
-	for _, sample := range samples {
-		sample := sample
-		// nolint:scopelint
-		t.Run(sample.name, func(t *testing.T) {
-			// t.Parallel()
-			for _, speed := range speeds {
-				speed := speed
-				// nolint:scopelint
-				t.Run(speed.name, func(t *testing.T) {
-					t.Parallel()
-					testMockSetSpeed(t, sample.firstSpeed*speed.multiplier, sample.secondSpeed*speed.multiplier)
-				})
-			}
-		})
-	}
-}
-
-func testMockSetSpeed(t *testing.T, firstSpeed, secondSpeed float64) {
-	t.Cleanup(func() { goleak.VerifyNone(t, goleak.IgnoreCurrent()) })
-
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-
-	const hours = 3
-
-	m := clockmock.New(ctx)
-
-	m.SetSpeed(firstSpeed)
-
-	realStart, mockStart := time.Now(), m.Now()
-	for i := 0; i < hours; i++ {
-		time.Sleep(testWait)
-		m.Add(time.Hour)
-	}
-	realDuration, mockDuration := time.Since(realStart), m.Since(mockStart)
-
-	m.SetSpeed(secondSpeed)
-
-	realStart, mockStart = time.Now(), m.Now()
-	for i := 0; i < hours; i++ {
-		time.Sleep(testWait)
-		m.Add(time.Hour)
-	}
-	realDuration += time.Since(realStart)
-	mockDuration += m.Since(mockStart)
-
-	mockSpeed := float64(mockDuration-2*hours*time.Hour) / float64(realDuration)
-	avgSpeed := (firstSpeed + secondSpeed) / 2
-
-	require.Greater(t, mockSpeed/avgSpeed, 0.6)
-	require.Less(t, mockSpeed/avgSpeed, 1.4)
-}
-
 func TestMock_Timer(t *testing.T) {
-	t.Parallel()
-	t.Cleanup(func() { goleak.VerifyNone(t, goleak.IgnoreCurrent()) })
+	t.Cleanup(func() { goleak.VerifyNone(t) })
 
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
@@ -183,8 +68,7 @@ func TestMock_Timer(t *testing.T) {
 }
 
 func TestMock_Timer_ZeroDuration(t *testing.T) {
-	t.Parallel()
-	t.Cleanup(func() { goleak.VerifyNone(t, goleak.IgnoreCurrent()) })
+	t.Cleanup(func() { goleak.VerifyNone(t) })
 
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
@@ -201,8 +85,7 @@ func TestMock_Timer_ZeroDuration(t *testing.T) {
 }
 
 func TestMock_Timer_Stop(t *testing.T) {
-	t.Parallel()
-	t.Cleanup(func() { goleak.VerifyNone(t, goleak.IgnoreCurrent()) })
+	t.Cleanup(func() { goleak.VerifyNone(t) })
 
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
@@ -224,8 +107,7 @@ func TestMock_Timer_Stop(t *testing.T) {
 }
 
 func TestMock_Timer_StopResult(t *testing.T) {
-	t.Parallel()
-	t.Cleanup(func() { goleak.VerifyNone(t, goleak.IgnoreCurrent()) })
+	t.Cleanup(func() { goleak.VerifyNone(t) })
 
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
@@ -241,8 +123,7 @@ func TestMock_Timer_StopResult(t *testing.T) {
 }
 
 func TestMock_Timer_Reset(t *testing.T) {
-	t.Parallel()
-	t.Cleanup(func() { goleak.VerifyNone(t, goleak.IgnoreCurrent()) })
+	t.Cleanup(func() { goleak.VerifyNone(t) })
 
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
@@ -274,8 +155,7 @@ func TestMock_Timer_Reset(t *testing.T) {
 }
 
 func TestMock_Timer_ResetExpired(t *testing.T) {
-	t.Parallel()
-	t.Cleanup(func() { goleak.VerifyNone(t, goleak.IgnoreCurrent()) })
+	t.Cleanup(func() { goleak.VerifyNone(t) })
 
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
@@ -308,8 +188,7 @@ func TestMock_Timer_ResetExpired(t *testing.T) {
 }
 
 func TestMock_Timer_Reset_ZeroDuration(t *testing.T) {
-	t.Parallel()
-	t.Cleanup(func() { goleak.VerifyNone(t, goleak.IgnoreCurrent()) })
+	t.Cleanup(func() { goleak.VerifyNone(t) })
 
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
@@ -329,8 +208,7 @@ func TestMock_Timer_Reset_ZeroDuration(t *testing.T) {
 }
 
 func TestMock_AfterFunc(t *testing.T) {
-	t.Parallel()
-	t.Cleanup(func() { goleak.VerifyNone(t, goleak.IgnoreCurrent()) })
+	t.Cleanup(func() { goleak.VerifyNone(t) })
 
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
@@ -362,8 +240,7 @@ func TestMock_AfterFunc(t *testing.T) {
 }
 
 func TestMock_AfterFunc_ZeroDuration(t *testing.T) {
-	t.Parallel()
-	t.Cleanup(func() { goleak.VerifyNone(t, goleak.IgnoreCurrent()) })
+	t.Cleanup(func() { goleak.VerifyNone(t) })
 
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
@@ -381,8 +258,7 @@ func TestMock_AfterFunc_ZeroDuration(t *testing.T) {
 }
 
 func TestMock_Ticker(t *testing.T) {
-	t.Parallel()
-	t.Cleanup(func() { goleak.VerifyNone(t, goleak.IgnoreCurrent()) })
+	t.Cleanup(func() { goleak.VerifyNone(t) })
 
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
@@ -417,8 +293,7 @@ func TestMock_Ticker(t *testing.T) {
 }
 
 func TestMock_WithDeadline(t *testing.T) {
-	t.Parallel()
-	t.Cleanup(func() { goleak.VerifyNone(t, goleak.IgnoreCurrent()) })
+	t.Cleanup(func() { goleak.VerifyNone(t) })
 
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
@@ -446,8 +321,7 @@ func TestMock_WithDeadline(t *testing.T) {
 }
 
 func TestMock_WithDeadline_Expired(t *testing.T) {
-	t.Parallel()
-	t.Cleanup(func() { goleak.VerifyNone(t, goleak.IgnoreCurrent()) })
+	t.Cleanup(func() { goleak.VerifyNone(t) })
 
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
@@ -466,8 +340,7 @@ func TestMock_WithDeadline_Expired(t *testing.T) {
 }
 
 func TestMock_WithDeadline_ParentCanceled(t *testing.T) {
-	t.Parallel()
-	t.Cleanup(func() { goleak.VerifyNone(t, goleak.IgnoreCurrent()) })
+	t.Cleanup(func() { goleak.VerifyNone(t) })
 
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
@@ -497,8 +370,7 @@ func TestMock_WithDeadline_ParentCanceled(t *testing.T) {
 }
 
 func TestMock_WithTimeout(t *testing.T) {
-	t.Parallel()
-	t.Cleanup(func() { goleak.VerifyNone(t, goleak.IgnoreCurrent()) })
+	t.Cleanup(func() { goleak.VerifyNone(t) })
 
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
